feat: Stream腾讯云部署 #7028 auth启动调整
